+++ tests/models/data/horovod/train_default_model.py
@@ -21,6 +21,7 @@ import json
 import os
 import sys
 
+
 try:
     import horovod.torch as hvd
 except (ModuleNotFoundError, ImportError):
@@ -33,7 +34,8 @@ sys.path.insert(0, os.path.abspath(PATH_ROOT))
 from pytorch_lightning import Trainer  # noqa: E402
 from pytorch_lightning.callbacks import ModelCheckpoint  # noqa: E402
 from tests.base import EvalModelTemplate  # noqa: E402
-from tests.base.develop_utils import set_random_master_port, run_model_test  # noqa: E402
+from tests.base.develop_pipelines import run_model_test  # noqa: E402
+from tests.base.develop_utils import set_random_master_port  # noqa: E402
 
 
 parser = argparse.ArgumentParser()
